pyproject.toml
[tool.poetry]
name = "snoopy-backend"
version = "0.1.0"
description = ""
authors = ["Bekhzod Tillakhanov <bekhzod.tillakhanov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
passlib = "^1.7.4"
SQLAlchemy = "^1.4.27"
alembic = "^1.7.5"
pydantic = "^1.8.2"
Flask = "^2.0.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
psycopg2-binary = "^2.9.2"
PyYAML = "^6.0"
pytest = "^6.2.5"
datamodel-code-generator = "^0.11.15"
click = "^8.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
